Rename jobs blueprint
