+++ blueprints/jobsBlueprint.ts
